# Feature/674 use properties type in filter