--- a/libs/error_injector/stdio/include/stdio_ei.h
+++ b/libs/error_injector/stdio/include/stdio_ei.h
@@ -44,6 +44,8 @@ CELIX_EI_DECLARE(fputc, int);
 
 CELIX_EI_DECLARE(fputs, int);
 
+CELIX_EI_DECLARE(fclose, int);
+
 #ifdef __cplusplus
 }
 #endif