--- a/libs/error_injector/stdio/src/stdio_ei.cc
+++ b/libs/error_injector/stdio/src/stdio_ei.cc
@@ -101,8 +101,11 @@ int __wrap_fputs(const char* __s, FILE* __stream) {
 int __real_fclose(FILE* __stream);
 CELIX_EI_DEFINE(fclose, int)
 int __wrap_fclose(FILE* __stream) {
+    int rc = __real_fclose(__stream); //note always call real fclose to ensure the stream is closed.
+    errno = ENOSPC;
     CELIX_EI_IMPL(fclose);
-    return __real_fclose(__stream);
+    errno = 0;
+    return rc;
 }
 
 }