--- a/libs/framework/src/dm_component_impl.c
+++ b/libs/framework/src/dm_component_impl.c
@@ -445,29 +445,42 @@ celix_status_t component_addInterface(celix_dm_component_t *component, const cha
     return celix_dmComponent_addInterface(component, serviceName, serviceVersion, service, properties);
 }
 
-celix_status_t celix_dmComponent_addInterface(celix_dm_component_t *component, const char* serviceName, const char* serviceVersion, const void* service, celix_properties_t* properties) {
+celix_status_t celix_dmComponent_addInterface(celix_dm_component_t* component,
+                                              const char* serviceName,
+                                              const char* serviceVersion,
+                                              const void* service,
+                                              celix_properties_t* properties) {
     if (serviceName == NULL || celix_utils_stringEquals(serviceName, "")) {
-        celix_bundleContext_log(component->context, CELIX_LOG_LEVEL_ERROR, "Cannot add interface with a NULL or empty serviceName");
+        celix_bundleContext_log(
+            component->context, CELIX_LOG_LEVEL_ERROR, "Cannot add interface with a NULL or empty serviceName");
         return CELIX_ILLEGAL_ARGUMENT;
     }
 
-    dm_interface_t *interface = calloc(1, sizeof(*interface));
-    char *name = celix_utils_strdup(serviceName);
+    dm_interface_t* interface = calloc(1, sizeof(*interface));
+    char* name = celix_utils_strdup(serviceName);
 
     if (properties == NULL) {
         properties = celix_properties_create();
     }
 
-    if ((properties_get(properties, CELIX_FRAMEWORK_SERVICE_VERSION) == NULL) && (serviceVersion != NULL)) {
-        celix_properties_set(properties, CELIX_FRAMEWORK_SERVICE_VERSION, serviceVersion);
+    if (serviceVersion != NULL) {
+        celix_version_t* version = celix_version_createVersionFromString(serviceVersion);
+        if (!version) {
+            celix_bundleContext_log(
+                component->context, CELIX_LOG_LEVEL_ERROR, "Cannot add interface with an invalid serviceVersion");
+            celix_properties_destroy(properties);
+            return CELIX_ILLEGAL_ARGUMENT;
+        }
+        celix_properties_setVersionWithoutCopy(properties, CELIX_FRAMEWORK_SERVICE_VERSION, version);
     }
+
     celix_properties_set(properties, CELIX_DM_COMPONENT_UUID, (char*)component->uuid);
 
     celixThreadMutex_lock(&component->mutex);
     interface->serviceName = name;
     interface->service = service;
     interface->properties = properties;
-    interface->svcId= -1L;
+    interface->svcId = -1L;
     celix_arrayList_add(component->providedInterfaces, interface);
     if (celix_dmComponent_currentState(component) == CELIX_DM_CMP_STATE_TRACKING_OPTIONAL) {
         celix_dmComponent_registerServices(component, false);