@pnoltes 2023-12-22
> 
> Please have a look at these two issues, @pnoltes
> .

Thanks, nice catch. I removed the incorrect whitespace removal